+++ tests/BeeMemoryBank.Integration.Tests/AclPathChangeTests.cs
@@ -74,7 +74,7 @@ public class AclPathChangeTests : IAsyncLifetime
         _folderSvc = new FolderService(_folderRepo, articleRepo, nodeRepo, clock, new NullEventLogger(), _folderAccessService);
 
         _httpContextAccessor = new HttpContextAccessor();
-        var responseManager = new McpResponseManager(Path.GetTempPath());
+        var responseManager = new McpResponseManager(Path.GetTempPath(), new HttpContextAccessor());
         var mediaService = new MediaService(mediaRepo, articleRepo, _session, nodeRepo, clock, new NullEventLogger(), mediaOptions);
         var copySvc = new CopyService(_articleService, _folderSvc, mediaService, articleRepo, _folderRepo, _conceptTagService, _scopeHolder);
         _writeTools = new BeeWriteTools(_articleService, _folderRepo, articleRepo, _folderSvc, copySvc, _conceptTagService, NullLogger<BeeWriteTools>.Instance, responseManager);